first arg to putc must be int or char, not long

+++ mach/proto/as/comm5.c
@@ -93,7 +93,7 @@ putval(c)
 		putc(c-128, tempfile);
 		v = yylval.y_valu;
 		while (--n >= 0)
-			putc(v >> (n*8), tempfile);
+			putc((int) (v >> (n*8)), tempfile);
 		return;
 	case IDENT:
 	case FBSYM:
